[REFACTOR]: Refactor common logic
