+++ src/phaser/scenes/MultiPlay/config.js
@@ -1,4 +1,5 @@
 import Phaser from "phaser";
+
 import Preloader from "./Preloader";
 import MultiPlayScene from "./MultiPlayScene";
